requirements.txt
Appium-Python-Client==1.2.0
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==10.1
selenium==3.141.0
toml==0.10.2
urllib3==1.26.6
